Faster 1.cr and 1.rb code for the lru benchmark
LCG generation in the Crystal code uses "&*" and "&+" operators
instead of "*" and "+" (two's complement wraparound instead of
arithmetic overflow exceptions). UInt64 is changed to UInt32,
because there's no reason to do slower 64-bit calculations.
LCG is now a lightweight "struct" instead of a heavyweight
"class".

Avoid doing redundant operations and make a better use of
"shift"/"delete" methods of the Hash class for both Crystal
and Ruby. This all results in a significant performance
improvement and smaller/cleaner code.