Tell Acton to target native platform
By default, Acton optimizes for portability, so on Linux, the output
binary targets GNU libc 2.27, which is rather old (~Ubuntu 18.04). It
also only uses base features of the x86_64 CPU, so features like AVX512
or similar would not be used. By setting --target native, the output
executable is instead optimized for the machine we are running on and
can typically run much faster, sometimes up to 2x faster.